Lives+coins UI (#33)
* Added new stage

Added a new stage for the lives mechanic

* Added Lives Canvas

* Added Lives Canvas

+ testEnemy and Storpingvin

* Damage and Health Reduction

Enemies now damage the Storpingvin when they collide. If the Storpingvin's lives are 0, it'll send a debug log with the message "Game Over".

* Added GameOver

Includes the function EndGame which ends the game. Function is called whenever the player's health <= 0. (not yet finished)

* Updated Health Reduction

* Pirat 1 Prefab

skapade och modifierade EnemyPrefabs.

* Added BaseEnemy, Enemymodel

Enemymodel ärver BaseEnemy.

* DealDamage

When the enemy reaches the end of the path, it now deals damage to the player. Removed Storpingvin (replaced by PlayerStats). EnemyModel now has variables for speed, damage and health. EnemyController has the DealDamage function in it.

* ScriptableObjects for EnemyStats

ScriptableObjects storing stats can now be created. Each enemy will need a scriptable object attached to their EnemyModel in order to contain stats.

* Removed unnecessary debug logs 

Cleaned up.

* (no branch): Auto stash before rebase of "HEAD"

* Prevented merge conflicts

Changed followpath directory and renamed "scripts".

* Merge branch 'feature/playerLives'

Conflicts:
	Penguin Tower Defense/Assets/Scrips/FollowPath.cs
	Penguin Tower Defense/Assets/Scripts/EnemyScripts/FollowPath.cs
	Penguin Tower Defense/Assets/TowerSeeking/TowerSeeking.cs.meta
	Penguin Tower Defense/Assets/game.meta
	Penguin Tower Defense/Assets/textures.meta
	Penguin Tower Defense/Assets/textures/enemies.meta
	Penguin Tower Defense/Assets/textures/enemies/Enemie 1.png.meta
	Penguin Tower Defense/Assets/textures/maps.meta
	Penguin Tower Defense/Assets/textures/maps/map 1.png.meta
	Penguin Tower Defense/Packages/manifest.json
	Penguin Tower Defense/UserSettings/EditorUserSettings.asset

* Ignorera user settings för att slippa konflikter.

* PlayerLives

Merged feature/playerLives into develop, added the necessary scripts to the prefabs. Also created "gameMaster", a game object where the player's HP (and possibly balance) is stored.

* LivesUI+canvas

The game now displays the amount of player lives on a canvas.

* Finalized the UI (lives + coins)

The game now displays the amount of player lives and coins on a canvas (level 1 onl). The canvas scales with screen size.

* Finalized the UI (lives + coins)

The game now displays the amount of player lives and coins on a canvas (level 1 only). The canvas scales with screen size.